+++ mod/bookmarks/views/default/resources/bookmarks/bookmarklet.php
@@ -3,7 +3,7 @@
 * Elgg bookmarks plugin bookmarklet page
 */
 
-$container_guid = elgg_extract('guid', $vars, elgg_extract('container_guid', $vars));
+$container_guid = (int) elgg_extract('guid', $vars, elgg_extract('container_guid', $vars));
 $container = get_entity($container_guid);
 $page_owner = $container;
