+++ mod/bookmarks/views/default/resources/bookmarks/edit.php
@@ -3,7 +3,7 @@
  * Add bookmark page
  */
 
-$bookmark_guid = elgg_extract('guid', $vars);
+$bookmark_guid = (int) elgg_extract('guid', $vars);
 elgg_entity_gatekeeper($bookmark_guid, 'object', 'bookmarks', true);
 
 $bookmark = get_entity($bookmark_guid);
