+++ mod/file/views/default/resources/file/group.php
@@ -3,7 +3,7 @@
  * Individual's or group's files
  */
 
-$group_guid = elgg_extract('guid', $vars);
+$group_guid = (int) elgg_extract('guid', $vars);
 
 elgg_entity_gatekeeper($group_guid, 'group');
