+++ mod/pages/views/default/resources/pages/view.php
@@ -5,7 +5,7 @@
 
 use Elgg\Exceptions\Http\EntityNotFoundException;
 
-$guid = elgg_extract('guid', $vars);
+$guid = (int) elgg_extract('guid', $vars);
 
 elgg_entity_gatekeeper($guid, 'object', 'page');
